+++ src/main/java/com/dantonov/musicstore/entity/Genre.java
@@ -12,6 +12,8 @@ import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
 import javax.persistence.OrderBy;
 import javax.persistence.Table;
+import org.hibernate.annotations.Cascade;
+import org.hibernate.annotations.CascadeType;
 
 /**
  *
@@ -25,7 +27,7 @@ public class Genre {
     @Id
     @Column(name = "genre_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Byte id;
+    private Integer id;
     
     @Column(name = "name", nullable = false, updatable = false, length = 32)
     private String name;
@@ -47,8 +49,8 @@ public class Genre {
 
     
     
-    public Byte getId() { return id; }
-    public void setId(Byte id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
