+++ src/main/java/com/dantonov/musicstore/entity/Role.java
@@ -11,6 +11,8 @@ import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
 import javax.persistence.Table;
+import org.hibernate.annotations.Cascade;
+import org.hibernate.annotations.CascadeType;
 
 /**
  *
@@ -24,11 +26,12 @@ public class Role {
     @Id
     @Column(name = "role_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Byte id;
+    private Integer id;
     
     @Column(name = "role", nullable = false, updatable = false, length = 16)
     private String role;
     
+
     @ManyToMany(targetEntity = User.class)
     @JoinTable(name = "Users_to_Roles",
                joinColumns = @JoinColumn(name = "role_id", nullable = false),
@@ -46,8 +49,8 @@ public class Role {
 
     
     
-    public Byte getId() { return id; }
-    public void setId(Byte id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getRole() { return role; }
     public void setRole(String role) { this.role = role; }
