+++ src/main/java/com/dantonov/musicstore/repository/GenreRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.CrudRepository;
  *
  * @author Denis Antonov (den007230@gmail.com)
  */
-public interface GenreRepository extends CrudRepository<Genre, Byte>{
+public interface GenreRepository extends CrudRepository<Genre, Integer>{
     
     Genre findByName(String name);
