+++ src/main/java/com/dantonov/musicstore/service/ActionService.java
@@ -19,7 +19,7 @@ public class ActionService {
     
     
     
-    public Action findById(Byte id) {
+    public Action findById(Integer id) {
         return actionRepository.findOne(id);
     }
